requirements_full.txt
mujoco
numpy
matplotlib.pyplot
tqdm
random
math
os
torch
multiprocessing